refactor: split ts node generator
