+++ codegen/ts_node_core.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+"""Public orchestration exports for the TypeScript Node SDK generator."""
+
+from ts_node_audio import gen_napi_rust_audio
+from ts_node_errors import gen_diagnostic, gen_errors
+from ts_node_interface import gen_input, gen_interface, gen_math
+from ts_node_napi import gen_napi_rust
+from ts_node_wrapper import gen_entry, gen_node_wrapper
