+++ examples/python/flappy_bird.py
@@ -70,6 +70,7 @@ class GameConstants:
 CONSTANTS = GameConstants()
 
 # Tiny embedded 8-bit PCM WAV clips (no external audio assets required).
+# Canonical decimal fixture bytes are documented in examples/shared/flappy_audio_fixture.txt.
 FLAP_WAV_BYTES = bytes([
     82, 73, 70, 70, 116, 0, 0, 0, 87, 65, 86, 69, 102, 109, 116, 32, 16, 0, 0, 0, 1, 0, 1, 0, 64, 31,
     0, 0, 64, 31, 0, 0, 1, 0, 8, 0, 100, 97, 116, 97, 80, 0, 0, 0, 127, 182, 191, 147, 87, 61, 88, 146,
